+++ packages/components/organisms/f-footer/test/specs/component/shared/f-footer.component.spec.js
@@ -21,7 +21,7 @@ describe('Shared - f-footer component tests', () => {
     });
 
     forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should not show courier links and country selector when options are unselected', expectedLocale => {
+    .it('should not show courier links and country selector for country code "%s" when options are unselected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -39,7 +39,7 @@ describe('Shared - f-footer component tests', () => {
     });
 
     forEach(['au', 'ie', 'nz'])
-    .it('should show courier links when option is selected', expectedLocale => {
+    .it('should show courier links for country code "%s" when option is selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -56,7 +56,7 @@ describe('Shared - f-footer component tests', () => {
     });
 
     forEach(['gb', 'es', 'it', 'no'])
-    .it('should never show courier links, even when option is selected', expectedLocale => {
+    .it('should never show courier links for country code "%s", even when option is selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -73,7 +73,7 @@ describe('Shared - f-footer component tests', () => {
     });
 
     forEach(['au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should always show country selector when selected', expectedLocale => {
+    .it('should always show country selector for country code "%s" when selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -90,7 +90,7 @@ describe('Shared - f-footer component tests', () => {
     });
 
     forEach(['au', 'gb', 'nz', 'ie', 'dk', 'es', 'it'])
-    .it('should display the corresponding icon for each locale selected', expectedLocale => {
+    .it('should display the corresponding country code ("%s") icon for each locale selected', expectedLocale => {
         // Arrange
         const footerData = {
             locale: expectedLocale,
@@ -106,9 +106,46 @@ describe('Shared - f-footer component tests', () => {
         expect(footer.isCurrentCountryIconDisplayed(expectedLocale)).toBe(true);
     });
 
-    forEach([['au', 'au'], ['at', 'at'], ['be', 'be-en'], ['bg', 'bg'], ['ca_en', 'skipthedishes.com'], ['ca_fr', 'skipthedishes.com/fr'], ['dk', '.dk'], ['jet_fr', '.fr'], ['de', '.de'], ['ie', '.ie'], ['il', '.il'], ['it', '.it'],
-        ['lu', 'lu-en'], ['nl', '.nl'], ['nz', '.nz'], ['no', '.no'], ['pl', '.pl'], ['pt', '/pt'], ['ro', '/ro'], ['es', '.es'], ['ch_ch', '.ch'], ['ch_en', '/en'], ['ch_fr', '/fr']])
-    .it('should display all countries and redirect to correct URL', (country, expectedUrl) => {
+    forEach([
+        ['au', 'au'],
+        ['at', 'at'],
+        ['no', 'no'],
+        ['dk', 'dk'],
+        ['be', 'be-en'],
+        ['bg', 'bg'],
+        ['ca_en', 'skipthedishes.com'],
+        ['jet_fr', '.fr'],
+        ['de', '.de'],
+        ['ie', '.ie'],
+        ['il', '.il'],
+        ['it', '.it']
+    ]).it('should display link for country code "%s" and redirect to correct URL', (country, expectedUrl) => { 
+        // Act
+        footer.clickCountrySelectorButton();
+        footer.expectedCountry = country;
+
+        // Assert
+        expect(footer.isCountryLinkItemDisplayed()).toBe(true);
+
+        // Act
+        footer.clickCountryLinkItem();
+
+        // Assert
+        expect(browser.getUrl()).toContain(expectedUrl);
+    });
+
+    forEach([
+        ['lu', 'lu-en'],
+        ['nl', '.nl'],
+        ['nz', '.nz'],
+        ['pl', '.pl'],
+        ['pt', '/pt'],
+        ['ro', '/ro'],
+        ['es', '.es'],
+        ['ch_ch', '.ch'],
+        ['ch_en', '/en'],
+        ['ch_fr', '/fr']])
+    .it('should display link for country code "%s" and redirect to correct URL', (country, expectedUrl) => {
         // Act
         footer.clickCountrySelectorButton();
         footer.expectedCountry = country;